Add quotes to yml strings and create .gitignore
This commit adds quotes to values in the services.yml file and creates a new .gitignore file. The quotes were added due to a change in how yml parses strings - without quotes, the parser was interpreting the keys incorrectly. The .gitignore file has been created to ignore IDEA files generated by the IDE.